Add TaskList class to handle task list operations
